move deep.py to nmf.py
