+++ torchnmf/__init__.py
@@ -5,6 +5,4 @@ __version__ = '0.2'
 __maintainer__ = 'Chin-Yun Yu'
 __email__ = 'lolimaster.cs03@nctu.edu.tw'
 
-from .deep import NMF, NMFD, NMF2D, NMF3D
-from .plca import PLCA, SIPLCA, SIPLCA2
-from .trainer import BetaMu, SparsityProj
+from . import nmf, plca, trainer, metrics
